refactor(android): Extract ViewModels into :viewmodel module; 2.16.27
All 5 screen-scoped ViewModels (Home/DoorHistory/Profile/FunctionList/
Diagnostics) + their Default impls and tests moved from :usecase to a
new :viewmodel KMP module. Two propagation tests that construct VMs
moved alongside.

Layer rule: :viewmodel may depend on :domain + :usecase only — no
:data, no :data-local, no concrete repository implementations. Source-
level enforcement via the existing checkLayerImports rule (the
.*ViewModel\.kt patterns now also cover :viewmodel/src/commonMain/
kotlin), build-level via the project() declarations in
viewmodel/build.gradle.kts.

Cleanup on :usecase: drop the androidx.lifecycle.viewmodel Gradle dep
+ the androidx.lifecycle. allowance in checkImportBoundary. :usecase
is now pure-Kotlin business logic with no Android lifecycle dep.
