+++ AndroidGarage/CHANGELOG.md
@@ -15,6 +15,13 @@ Internal release history. For Play Store "What's New" text, see `distribution/wh
 
 Every version gets an entry in this file (internal history). Play Store `distribution/whatsnew/` gets a line per minor/major — patches roll up into the next minor's line, or get a combined line if promoted to production on their own.
 
+## 2.16.27
+- **Internal: ViewModels extracted into a new `:viewmodel` Gradle module.** All 5 screen-scoped ViewModels (`HomeViewModel`, `DoorHistoryViewModel`, `ProfileViewModel`, `FunctionListViewModel`, `DiagnosticsViewModel`) plus their `Default*` impls and tests moved from `:usecase/src/commonMain/kotlin/com/chriscartland/garage/usecase/` to `:viewmodel/src/commonMain/kotlin/com/chriscartland/garage/viewmodel/`. Same KMP layout as `:usecase`. Two propagation tests (`RealNetworkButtonHealthRepositoryPropagationTest`, `RealNetworkSnoozeRepositoryPropagationTest`) moved alongside since they construct VM instances. No production-runtime change for users.
+- **Internal: `:viewmodel` allowed dependencies — `:domain` + `:usecase` only** (production / `commonMain`). Tests additionally depend on `:test-common` (fakes) and `:data` (real network repos for propagation tests). `:viewmodel` MUST NOT depend on `:data`, `:data-local`, or any concrete repository implementation — VMs orchestrate UI state via UseCases, never reach into the repository layer directly. Source-level enforcement: the `.*ViewModel\.kt` rules in `checkLayerImports` already block `data.*` / `domain.repository.*` imports, and the `viewmodel/src/commonMain/kotlin` source dir is now in every architecture-lint source-dir list (8 lists in root `build.gradle.kts`). Build-level: documented at the top of `viewmodel/build.gradle.kts`.
+- **Internal: `checkScreenViewModelCardinality` import filter updated** from `com.chriscartland.garage.usecase.` → `com.chriscartland.garage.viewmodel.`. The rule (one VM per `*Content.kt` screen) is unchanged; only the path it scans for changed.
+- **Internal: `:usecase` cleanup** — dropped the `androidx.lifecycle.viewmodel` Gradle dep and the `androidx.lifecycle.` allowance in its `checkImportBoundary` task. `:usecase` is now pure-Kotlin business logic with no Android lifecycle dependency, matching ADR intent.
+- **Internal: 6 androidApp consumer files** (`AppComponent.kt`, `HomeContent.kt`, `DoorHistoryContent.kt`, `ProfileContent.kt`, `FunctionListContent.kt`, `DiagnosticsContent.kt`) updated their `import com.chriscartland.garage.usecase.*ViewModel` paths to `com.chriscartland.garage.viewmodel.*ViewModel`. `AppComponent` now `implementation(project(":viewmodel"))` alongside the existing `:usecase` dep.
+
 ## 2.16.26
 - **Settings → Developer → Nav rail consolidated into one row + bottom sheet.** Pre-2.16.26 the Developer section had two inline controls for the rail (item-position picker + top-padding stepper) totalling ~3 visible rows. The Developer menu was getting long. Now there's a single `Nav rail` row whose subtitle shows the current state (e.g. `Top-aligned · 8 dp`); tap opens a `NavRailBottomSheet` containing both controls. Each control has its own consistent reset (`↺`) icon in its section header, so resetting one doesn't disturb the other. Pattern follows the existing `Snooze` / `Account` / `Version` bottom sheets — no new screen/route added, no navigation overhead.
 - **Default item position is now `Top-aligned`** (was `Centered vertically`). Combined with the existing 8 dp default for top padding (shipped 2.16.25), this means the rail-vs-content alignment is correct out of the box for new installs — the selected-item indicator pill (the rounded blue background drawn behind the icon when selected) lands at the same y as the body's first content row's top. Existing users who had explicitly set the value in 2.16.23/2.16.24 keep their override (DataStore stores absent-vs-set distinctly).
