+++ AndroidGarage/settings.gradle.kts
@@ -47,3 +47,4 @@ include(":presentation-model")
 include(":shared")
 include(":test-common")
 include(":usecase")
+include(":viewmodel")
