+++ AndroidGarage/version.properties
@@ -1,1 +1,1 @@
-versionName=2.16.26
+versionName=2.16.27
